* Corrected type definition of         argument passed to Tcl_GetStringFromObj() from size_t to int.         Incorrect type broke [pack] and [grid] on systems where         sizeof(size_t) != sizeof(int).  [Bugs 462375, 462342, 462338]

+++ tk/generic/tkGrid.c
@@ -2384,7 +2384,7 @@ ConfigureSlaves(interp, tkwin, objc, objv)
     Gridder *slavePtr;
     Tk_Window other, slave, parent, ancestor;
     int i, j, tmp;
-    size_t length;
+    int length;
     int numWindows;
     int width;
     int defaultColumn = 0;	/* default column number */
